商城业务-首页-整合thymeleaf渲染首页页面

+++ gulimall-ware/src/main/java/com/fosss/gulimall/ware/dao/WareSkuDao.java
@@ -15,5 +15,5 @@ import org.apache.ibatis.annotations.Param;
 @Mapper
 public interface WareSkuDao extends BaseMapper<WareSkuEntity> {
     //查询该sku的库存量
-    int hasStock(@Param("skuId") Long skuId);
+    Long hasStock(@Param("skuId") Long skuId);
 }
